Extract Repl and Guide
