+++ src/stores/drawerOpen.js
@@ -1,13 +1,9 @@
 import { writable } from 'svelte/store';
 
-function createDrawerOpen() {
-	const { subscribe, set, update } = writable(false);
+const { subscribe, set, update } = writable(false);
 
-	return {
-		subscribe,
-		toggle: () => update(open => !open),
-		close: () => set(false)
-	};
-}
-
-export default createDrawerOpen();
+export default {
+	subscribe,
+	toggle: () => update(open => !open),
+	close: () => set(false)
+};
